Refactor: 스킬 파일 리펙토링
- 스킬 파일 리펙토링
- 파일 분리
