+++ Assets/Workers/DEV/HJS/Scripts/Interface/ISpec.cs
@@ -9,5 +9,5 @@ public interface ISpec
     /// </summary>
     /// <param name="spec">정보</param>
     /// <param name="level">레벨</param>
-    public void SetSpec(BaseSkillSO.Spec spec, int level);
+    public void SetSpec(Spec spec, int level);
 }
