+++ Assets/Workers/DEV/HJS/Scripts/Skills/Base/ActiveSkill.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 29f84f93039f46d4c8faf8ad9d7a9040
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
