+++ Assets/Workers/DEV/HJS/Scripts/Skills/Base/PassiveSkill.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 281ffbbbf4d5e324e821bdec94019597
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
