+++ Assets/Workers/DEV/HJS/Scripts/Skills/Base/Spec.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d8c22c6cfecf17b42aee6b8ea9b1f744
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
